Clean up RuleOrdering
Previously rule orderings were split across all the different
rules. With this change, I have introduced a new RuleOrder static
class that has constants for each rule sorted by the order in which
they run.  This will make it easier to see the ordering of rules and
make additions of new rules easier to review.

Perhaps longer term we'll want to move to a system where rules declare
dependencies or something and we build an ordering based on them, but
for now I think this is better than the existing pattern.