Skip non-existent and read-only files
Previously, if a project included file paths which did not actually
exist, Workspace.TryApplyChanges() would create empty files at those
paths.  Also, if a file was read-only, the entire formatting operation
would fail with an UnauthorizedAccessException.  This change makes sure
we skip the document in these cases and emit a warning instead.