Skip non-existent and read-only files
Previously, if a project included file paths which did not actually
exist, Workspace.TryApplyChanges() would create empty files at those
paths. Also, if a file was read-only, the entire formatting operation
would fail with an UnauthorizedAccessException. This change makes sure
we skip the document in these cases and emit a warning instead.

This change moves away from using Workspace.TryApplyChanges() because
that method attempts to apply *all* workspace changes, including file
additions, removals, etc. Since this tool just produces clean Document
objects as output--that is, we don't want to modify any of the other
files that make up the workspace--it makes sense to write the document
text ourselves.