Passing a project should only use that project
When passing a project file name the formatter was still using the
resulting Workspace to drive the formatter.  In the case a project is a
part of a larger solution loading a project will load the larger
solution, not the individual project.  This means that every possible
document in the solution was processed instead of the project.

Changed the formatter to process the project only in that case.