Delay retrieving the SemanticModel element
Getting the SemanticModel instance is potentially an expensive
operation.  For example in the Roslyn.sln project it takes on average
3-5 seconds if it requires creating a new Compilation.  This is true
even when processing a single project because MSBuild will load the
entire solution for such a request and hence the Compilation encompasses
the entire solution.

Delaying access in this manner significantly speeds up the formatting
speed of documents that need no changes.