Basic structure working
The new code is actually so fast that it is hard to keep up with the
output.  It can run very quickly on the CodeAnalysis project now except
there is a bug in private field rename.  Need to cleanup the output to
be useable again.  Then onto debugging that issue.