Allow copyright header customization
This change adds internal support for customizing the copyright header
the formatter adds.  This still needs to get an official command line
option but that can be added later.

The CombinationTests suite now passes again as well.  When the test was
first written the copyright rule was disabled as Roslyn used a different
copyright.  When I re-enabled that rule I forgot to run the test and
hence it ended up in a broken state.