Abstracted out writing to the console
Added the IFormatLogger interface to abstract away the direct access to
the Console type.  This is a library that can be hosted in a variety of
environments hence it can't write directly to the console.  It needs an
abstraction the host can use to decide how to display the output.

This also made the unit test output clean again.