Add method to exempt code from formatting
This change adds the preprocessor symbol DOTNET_FORMATTER to every
project in the formatted Solution.  This allows developers to exempt
code from formatting by surrounding blocks with

``` csharp
#if !DOTNET_FORMATTER
```

This should only be done on the rare occasions where insignificant white
space serves to make the code significantly more readable.  For example
in large unicode tables.  In general this feature should be avoided.

closes #37