Remove simplification rule
This rule was not currently running due to the performance implications and unexpected churn.  Additionally it was attempting to simplify all constructs vs. just those which are called out in the coding guidelines.  Remove it in favor of going forward with ExplicitThisRule which is essentially a targetted version of this rule.