Ensure Sys.Col.Immutable copied to output
The code formatter depends on System.Collections.Immutable at runtime to
operate.  This is a value which is not guaranteed to be in the GAC and
hence must be deployed with the tool to ensure it runs correctly.

The easiest way to do this is to ensure it is a part of the build
output.  That way the output directory can be zipped up and deployed
without the need for custom actions.

Unfortunately other tools installed on developers machines were putting
the exact version of System.Collections.Immutable used by the tool into
the GAC.  This caused the MSBuild logic to fail to copy the reference to
the output even if you specified CopyLocal=true.  Making it Private=true
overrides even the GAC exception and the DLL does get copied.

closes #45