Don't move usings inside #if directives
This changes the move usings outside of a namespace rule to be very wary
of #if directives.  There is simply no safe way to move a directive of
the following nature:

``` csharp
namespace NS1
{
#if COND
using NS2;
class C { }
#else
using NS3;
class C { }
#endif
```

Lacking a good API to detect a using inside a #if directive we simply
abandon this rule if there are any #if directives in the file.

closes #71