Test for multiple namespaces
Add a test that verifies we don't do any moves when there are multiple
namespaces in the file.  There is no way to safely move a using in this
case using only syntax information.

This move effectively adds new namespaces at a global level which used
to be local to a namespace.  This means extension methods and types from
this namespace will affect compilation in all namespaces.  This can
introduce both compilation errors or even worse silently change overload
resolution in the file.