Fix VB trivia handling with comments + attrs
The explicit modifier code for VB wasn't correctly handling the leading
trivia on types and methods.  It mistakenly grabbed and replaced the
trivia on the node instead of the appropriate keyword token when there
were no other modifiers.  This resulted in strange effects like comments
being moved below attributes.

Fixed the first modifier case to use the keyword for trivial handling
instead.