Disable table formatting by default
This changes the behavior of the formatter to not define
DOTNET_FORMATTER by default.  It can be opted into by using the /tables
option.

The reason for this change is to avoid a bug in Roslyn around changing
and resetting the parser options.  When hit Roslyn will end up changing
all files in the solution to have a different encoding / line ending.

I haven't yet tracked down this bug but short term making this opt-in
behavior so users aren't adversely affected by it.