Update command line parsing
The command line parsing code didn't reflect the current state of the
tool.  The <rule types> switch for instance was only used to add in the
xUnit conversion rules which are no longer a part of the code style
tool.   This and some other dead code was removed and it now reflects
the real feature set of the tool.