Add DeadRegions tool projects to solution
Also updates the rest of the projects to use the common build.props
file in the repository root.

Strange thing to note is: I started getting a 'CS1985: Cannot await in
the body of a catch block' error in AnalysisEngine.RegionRemoval.cs but
not consistently. So I've fixed that line.