Explicit visibility trivia handling
This change addresses how we handle trivia when inserting explicit
visibility on a member / type.  The previous code incorrectly used
the leading trivia of the node in question to position the first
visibility modifier.  That would end up combining trivia from say
before and after an attribute attached to the node which was
incorrect.  Now we grab the trivia from the node / type itself
and use that.

closes #86