Add a switch to opt-out of slow checks
Some checks that we will soon implement are rather slow
and would prevent using the tool as a normal part of commits.
Create a switch so that slower tests can be opt-in, so that the
default (or simple) experience is fast and simple, so it can be run
frequently.

Fix #111