Verify Assert Argument Order
Test assert methods assume that the "actual" value is a variable
and the "expected" value is a constant for the most part. Inverting
these arguments causes misleading test failure messages that hinder
the ability to quickly diagnose test failures.

Detecting obvious mistakes (where "actual" is constant, and
"expected" is variable) can be done and the arguments swapped.

Fix #105