Implement FormatProjectWithAnalyzersAsync
This is the guts of the analyzer-based implementation. But we can't actually enable it yet, because we haven't yet written any analyzers, so the required call to compilation.WithAnalyzers would fail.

We are only making this commit to ensure that we have a convenient point where all the rule-based code still works.