MEF 2 (cont'd): Reimplement FormattingEngine.GetFormattingRules
In the previous commit, we commented out the body of FormattingEngine.GetFormattingRules, because in MEF 2, the object returned from GetExports does not expose a Metadata property.

But IFormattingEngine already exposes a property AllRules, so the easiest thing to do is just create an engine and ask it for the rules.

This still won't work, because as of this commit the metadata type is still IRuleMetadata, and MEF 2 can't use an abstract type as its metadata view. We'll work on that next.