MEF 2 (cont'd): Fix the fixing
The default values for the metadata properties were not being correctly initialized. As a result, all the rules were turned off. Fix the initialization. Now the code fixing actually works.

The same 8 unit tests still fail as in the last commit, but now the error is "Copyright.md not found". We'll fix that in the next commit.