MEF 2 (concluded): Fix broken unit tests
Some tests were broken because of the fact that in MEF 2, the default part creation policy is "non-shared", whereas in MEF 1 it was "shared".

These tests (in CombinationTest.cs) first created a FormattingEngineObject. Since that class's importing constructor takes an Options object, MEF creates one and passes it to the constructor. These tests then modified the Options object, for example, by setting its CopyrightHeader property.

Next, these tests would tell the FormattingEngineObject to perform a format. This involved instantiating all the formatting rule objects. Each formatting rule class has an importing constructor that takes an Options object.

Now in MEF 1, where the default part creation policy was "shared", each rule class was instantiated with the existing Options object. But in MEF 2, where the default policy is non-shared, each rule was instantiated with a *new* Options object -- and *not* the one whose properties had been set up for purposes of the test.