Fix double new lines with extern directives
Fix the double new lines which got inserted in the presence of an extern
directive in the code base.

Note: As a part of this change I had to completely revisit the new line
above rule because it didn't take into account a number of subtle
scenarios.  The rules around #pragma were simply making the
implementation very difficult to get correct and hence I decided to
change them to make it more managable.

closes #83