Implement UnwritableFieldFixer
Also:
o Fix bug in UnwritableFieldAnalyzer: it was creating a Diagnostic but not actually reproting it.

Note: The analyzer still treats every field as unwritten, so the fixer will mark every field with readonly.

Now first set of UnwritableFieldAnalyerTests pass, but the ExplicitThisAnalyzerTests fail, because as the code stands, the unit tests run all the rules, not just the one being tested -- and so the output of the ExplicitThisAnalyzerTests includes some readonly modifiers. So the next step will be to add an option to the formatting engine to allow a set of analyzers to be specified.